+++ crates/world/flashblocks/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod args;
 pub mod builder;
-pub mod finish_flashblock;
 pub mod payload;
 pub mod payload_builder_ctx;
