+++ crates/world/node/Cargo.toml
@@ -16,6 +16,8 @@ world-chain-builder-pool.workspace = true
 world-chain-builder-test-utils = { workspace = true, optional = true }
 world-chain-builder-pbh = { workspace = true, optional = true }
 
+flashblocks.workspace = true
+
 reth.workspace = true
 reth-chain-state = { workspace = true, optional = true }
 reth-db = { workspace = true, optional = true }
@@ -64,27 +66,27 @@ alloy-signer-local.workspace = true
 alloy-eips.workspace = true
 
 criterion.workspace = true
-semaphore-rs = { workspace = true, features = ["depth_30"]}
+semaphore-rs = { workspace = true, features = ["depth_30"] }
 serde_json.workspace = true
 chrono.workspace = true
 
 [features]
 default = []
 test = [
-    "world-chain-builder-test-utils",
-    "world-chain-builder-pbh",
-    "dep:reth-e2e-test-utils",
-    "dep:reth-chain-state",
-    "dep:reth-primitives",
-    "dep:reth-primitives-traits",
-    "dep:reth-prune-types",
-    "dep:reth-trie",
-    "dep:reth-db",
-    "dep:alloy-rpc-types",
-    "dep:alloy-sol-types",
-    "dep:alloy-eips",
-    "dep:revm-primitives",
-    "dep:futures",
-    "dep:chrono",
-    "dep:tokio",
+  "world-chain-builder-test-utils",
+  "world-chain-builder-pbh",
+  "dep:reth-e2e-test-utils",
+  "dep:reth-chain-state",
+  "dep:reth-primitives",
+  "dep:reth-primitives-traits",
+  "dep:reth-prune-types",
+  "dep:reth-trie",
+  "dep:reth-db",
+  "dep:alloy-rpc-types",
+  "dep:alloy-sol-types",
+  "dep:alloy-eips",
+  "dep:revm-primitives",
+  "dep:futures",
+  "dep:chrono",
+  "dep:tokio",
 ]
