+++ crates/world/payload/Cargo.toml
@@ -15,6 +15,8 @@ workspace = true
 world-chain-builder-pool.workspace = true
 world-chain-builder-rpc.workspace = true
 
+flashblocks.workspace = true
+
 # reth
 reth.workspace = true
 reth-basic-payload-builder.workspace = true
@@ -24,6 +26,7 @@ reth-optimism-node.workspace = true
 reth-optimism-chainspec.workspace = true
 reth-optimism-payload-builder.workspace = true
 reth-optimism-primitives.workspace = true
+reth-optimism-forks.workspace = true
 reth-provider.workspace = true
 reth-primitives.workspace = true
 reth-transaction-pool.workspace = true
