+++ crates/world/payload/src/builder.rs
@@ -1,17 +1,11 @@
-use alloy_consensus::{SignableTransaction, Transaction};
-use alloy_eips::Typed2718;
-use alloy_network::{TransactionBuilder, TxSignerSync};
-use alloy_rlp::Encodable;
 use alloy_rpc_types_debug::ExecutionWitness;
 use alloy_signer_local::PrivateKeySigner;
-use eyre::eyre::eyre;
-use op_alloy_rpc_types::OpTransactionRequest;
-use op_revm::OpContext;
+use flashblocks::payload_builder_ctx::PayloadBuilderCtx;
 use reth::api::PayloadBuilderError;
 use reth::payload::PayloadBuilderAttributes;
 use reth::revm::database::StateProviderDatabase;
 use reth::revm::witness::ExecutionWitnessRecord;
-use reth::revm::{DatabaseCommit, State};
+use reth::revm::State;
 use reth::transaction_pool::{BestTransactionsAttributes, TransactionPool};
 use reth_basic_payload_builder::{
     BuildArguments, BuildOutcome, BuildOutcomeKind, MissingPayloadBehaviour, PayloadBuilder,
@@ -19,41 +13,31 @@ use reth_basic_payload_builder::{
 };
 use reth_chain_state::{ExecutedBlock, ExecutedBlockWithTrieUpdates};
 use reth_evm::execute::BlockBuilderOutcome;
-use reth_evm::execute::BlockExecutionError;
-use reth_evm::execute::BlockValidationError;
 use reth_evm::execute::{BlockBuilder, BlockExecutor};
+use reth_evm::Database;
 use reth_evm::Evm;
-use reth_evm::{ConfigureEvm, Database};
 use reth_optimism_chainspec::OpChainSpec;
 use reth_optimism_node::{
-    OpBuiltPayload, OpEvm, OpEvmConfig, OpNextBlockEnvAttributes, OpPayloadBuilder,
-    OpPayloadBuilderAttributes,
+    OpBuiltPayload, OpEvmConfig, OpPayloadBuilder, OpPayloadBuilderAttributes,
 };
-use reth_optimism_payload_builder::builder::{
-    ExecutionInfo, OpPayloadBuilderCtx, OpPayloadTransactions,
-};
+use reth_optimism_payload_builder::builder::{OpPayloadBuilderCtx, OpPayloadTransactions};
 use reth_optimism_payload_builder::config::OpBuilderConfig;
 use reth_optimism_payload_builder::OpPayloadAttributes;
 use reth_optimism_primitives::{OpPrimitives, OpTransactionSigned};
 use reth_payload_util::{NoopPayloadTransactions, PayloadTransactions};
-use reth_primitives::{Block, Recovered, SealedHeader};
-use reth_primitives_traits::SignedTransaction;
+use reth_primitives::{Block, SealedHeader};
 use reth_provider::{
     BlockReaderIdExt, ChainSpecProvider, ExecutionOutcome, ProviderError, StateProvider,
     StateProviderFactory,
 };
-use reth_transaction_pool::{BlobStore, PoolTransaction};
-use revm::inspector::NoOpInspector;
-use revm::Inspector;
-use revm_primitives::{Address, U256};
-use semaphore_rs::Field;
-use std::collections::HashSet;
+use reth_transaction_pool::BlobStore;
+use revm_primitives::Address;
 use std::sync::Arc;
-use tracing::{debug, error, trace};
-use world_chain_builder_pool::bindings::IPBHEntryPoint::spendNullifierHashesCall;
+use tracing::debug;
 use world_chain_builder_pool::tx::{WorldChainPoolTransaction, WorldChainPooledTransaction};
 use world_chain_builder_pool::WorldChainTransactionPool;
-use world_chain_builder_rpc::transactions::validate_conditional_options;
+
+use crate::ctx::WorldChainPayloadBuilderCtx;
 
 /// World Chain payload builder
 #[derive(Debug, Clone)]
@@ -231,6 +215,7 @@ where
             pbh_entry_point: self.pbh_entry_point,
             pbh_signature_aggregator: self.pbh_signature_aggregator,
             builder_private_key: self.builder_private_key.clone(),
+            pool: self.inner.pool.clone(),
         };
 
         let op_ctx = &ctx.inner;
@@ -284,6 +269,7 @@ where
             pbh_entry_point: self.pbh_entry_point,
             pbh_signature_aggregator: self.pbh_signature_aggregator,
             builder_private_key: self.builder_private_key.clone(),
+            pool: self.inner.pool.clone(),
         };
 
         let state_provider = self
@@ -294,7 +280,7 @@ where
         let builder: WorldChainBuilder<'_, NoopPayloadTransactions<WorldChainPooledTransaction>> =
             WorldChainBuilder::new(|_| NoopPayloadTransactions::default());
 
-        builder.witness(state_provider, &ctx, &self.inner.pool)
+        builder.witness(state_provider, &ctx)
     }
 }
 
@@ -386,7 +372,7 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
         self,
         db: impl Database<Error = ProviderError>,
         state_provider: impl StateProvider,
-        ctx: WorldChainPayloadBuilderCtx<Client>,
+        ctx: WorldChainPayloadBuilderCtx<Client, Pool>,
         pool: &Pool,
     ) -> Result<BuildOutcomeKind<OpBuiltPayload<OpPrimitives>>, PayloadBuilderError>
     where
@@ -412,7 +398,7 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
         debug!(target: "payload_builder", id=%op_ctx.payload_id(), parent_header = ?ctx.inner.parent().hash(), parent_number = ctx.inner.parent().number, "building new payload");
 
         // Prepare block builder.
-        let mut builder = ctx.block_builder(&mut state)?;
+        let mut builder = PayloadBuilderCtx::block_builder(&ctx, &mut state)?;
 
         // 1. apply pre-execution changes
         builder.apply_pre_execution_changes()?;
@@ -423,8 +409,9 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
         // 3. if mem pool transactions are requested we execute them
         if !op_ctx.attributes().no_tx_pool {
             let best_txs = best(op_ctx.best_transaction_attributes(builder.evm_mut().block()));
+            // TODO: Validate gas limit
             if ctx
-                .execute_best_transactions(&mut info, &mut builder, best_txs, pool)?
+                .execute_best_transactions(&mut info, &mut builder, best_txs, 0)?
                 .is_some()
             {
                 return Ok(BuildOutcomeKind::Cancelled);
@@ -489,8 +476,7 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
     pub fn witness<Pool, Client>(
         self,
         state_provider: impl StateProvider,
-        ctx: &WorldChainPayloadBuilderCtx<Client>,
-        pool: &Pool,
+        ctx: &WorldChainPayloadBuilderCtx<Client, Pool>,
     ) -> Result<ExecutionWitness, PayloadBuilderError>
     where
         Txs: PayloadTransactions<
@@ -510,7 +496,7 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
             .with_database(StateProviderDatabase::new(&state_provider))
             .with_bundle_update()
             .build();
-        let mut builder = ctx.block_builder(&mut db)?;
+        let mut builder = PayloadBuilderCtx::block_builder(ctx, &mut db)?;
 
         builder.apply_pre_execution_changes()?;
         let mut info = ctx.inner.execute_sequencer_transactions(&mut builder)?;
@@ -519,7 +505,8 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
                 ctx.inner
                     .best_transaction_attributes(builder.evm_mut().block()),
             );
-            ctx.execute_best_transactions(&mut info, &mut builder, best_txs, pool)?;
+            // TODO: Validate gas limit
+            ctx.execute_best_transactions(&mut info, &mut builder, best_txs, 0)?;
         }
         builder.into_executor().apply_post_execution_changes()?;
 
@@ -535,285 +522,4 @@ impl<Txs> WorldChainBuilder<'_, Txs> {
             keys,
         })
     }
-}
-
-/// Container type that holds all necessities to build a new payload.
-#[derive(Debug)]
-pub struct WorldChainPayloadBuilderCtx<Client> {
-    pub inner: OpPayloadBuilderCtx<OpEvmConfig, OpChainSpec>,
-    pub verified_blockspace_capacity: u8,
-    pub pbh_entry_point: Address,
-    pub pbh_signature_aggregator: Address,
-    pub client: Client,
-    pub builder_private_key: PrivateKeySigner,
-}
-
-impl<Client> WorldChainPayloadBuilderCtx<Client>
-where
-    Client: StateProviderFactory
-        + BlockReaderIdExt<Block = Block<OpTransactionSigned>>
-        + ChainSpecProvider<ChainSpec = OpChainSpec>
-        + Clone,
-{
-    /// Executes the given best transactions and updates the execution info.
-    ///
-    /// Returns `Ok(Some(())` if the job was cancelled.
-    pub fn execute_best_transactions<TXS, DB, Builder, Pool>(
-        &self,
-        info: &mut ExecutionInfo,
-        builder: &mut Builder,
-        mut best_txs: TXS,
-        pool: &Pool,
-    ) -> Result<Option<()>, PayloadBuilderError>
-    where
-        DB: Database + DatabaseCommit,
-        Builder: BlockBuilder<
-            Primitives = OpPrimitives,
-            Executor: BlockExecutor<Evm = OpEvm<DB, NoOpInspector>>,
-        >,
-        TXS: PayloadTransactions<
-            Transaction: WorldChainPoolTransaction<Consensus = OpTransactionSigned>,
-        >,
-        Pool: TransactionPool<
-            Transaction: WorldChainPoolTransaction<Consensus = OpTransactionSigned>,
-        >,
-    {
-        let mut block_gas_limit = builder.evm_mut().block().gas_limit;
-        let block_da_limit = self.inner.da_config.max_da_block_size();
-        let tx_da_limit = self.inner.da_config.max_da_tx_size();
-        let base_fee = builder.evm_mut().block().basefee;
-
-        let mut invalid_txs = vec![];
-        let verified_gas_limit = (self.verified_blockspace_capacity as u64 * block_gas_limit) / 100;
-
-        let mut spent_nullifier_hashes = HashSet::new();
-        while let Some(pooled_tx) = best_txs.next(()) {
-            let tx = pooled_tx.clone().into_consensus();
-            if info.is_tx_over_limits(tx.inner(), block_gas_limit, tx_da_limit, block_da_limit) {
-                // we can't fit this transaction into the block, so we need to mark it as
-                // invalid which also removes all dependent transaction from
-                // the iterator before we can continue
-                best_txs.mark_invalid(tx.signer(), tx.nonce());
-                continue;
-            }
-
-            if let Some(conditional_options) = pooled_tx.conditional_options() {
-                if validate_conditional_options(conditional_options, &self.client).is_err() {
-                    best_txs.mark_invalid(tx.signer(), tx.nonce());
-                    invalid_txs.push(*pooled_tx.hash());
-                    continue;
-                }
-            }
-
-            // ensure we still have capacity for this transaction
-            if info.cumulative_gas_used + tx.gas_limit() > block_gas_limit {
-                // we can't fit this transaction into the block, so we need to mark it as
-                // invalid which also removes all dependent transaction from
-                // the iterator before we can continue
-                best_txs.mark_invalid(tx.signer(), tx.nonce());
-                continue;
-            }
-
-            // A sequencer's block should never contain blob or deposit transactions from the pool.
-            if tx.is_eip4844() || tx.is_deposit() {
-                best_txs.mark_invalid(tx.signer(), tx.nonce());
-                continue;
-            }
-
-            // check if the job was cancelled, if so we can exit early
-            if self.inner.cancel.is_cancelled() {
-                return Ok(Some(()));
-            }
-
-            // If the transaction is verified, check if it can be added within the verified gas limit
-            if let Some(payloads) = pooled_tx.pbh_payload() {
-                if info.cumulative_gas_used + tx.gas_limit() > verified_gas_limit {
-                    best_txs.mark_invalid(tx.signer(), tx.nonce());
-                    continue;
-                }
-
-                if payloads
-                    .iter()
-                    .any(|payload| !spent_nullifier_hashes.insert(payload.nullifier_hash))
-                {
-                    best_txs.mark_invalid(tx.signer(), tx.nonce());
-                    invalid_txs.push(*pooled_tx.hash());
-                    continue;
-                }
-            }
-
-            let gas_used = match builder.execute_transaction(tx.clone()) {
-                Ok(res) => {
-                    if let Some(payloads) = pooled_tx.pbh_payload() {
-                        if spent_nullifier_hashes.len() == payloads.len() {
-                            block_gas_limit -= FIXED_GAS
-                        }
-
-                        block_gas_limit -= COLD_SSTORE_GAS * payloads.len() as u64;
-                    }
-                    res
-                }
-                Err(err) => {
-                    match err {
-                        BlockExecutionError::Validation(BlockValidationError::InvalidTx {
-                            error,
-                            ..
-                        }) => {
-                            if error.is_nonce_too_low() {
-                                // if the nonce is too low, we can skip this transaction
-                                trace!(target: "payload_builder", %error, ?tx, "skipping nonce too low transaction");
-                            } else {
-                                // if the transaction is invalid, we can skip it and all of its
-                                // descendants
-                                trace!(target: "payload_builder", %error, ?tx, "skipping invalid transaction and its descendants");
-                                best_txs.mark_invalid(tx.signer(), tx.nonce());
-                            }
-
-                            continue;
-                        }
-
-                        err => {
-                            // this is an error that we should treat as fatal for this attempt
-                            return Err(PayloadBuilderError::EvmExecutionError(Box::new(err)));
-                        }
-                    }
-                }
-            };
-
-            self.commit_changes(info, base_fee, gas_used, tx);
-        }
-
-        if !spent_nullifier_hashes.is_empty() {
-            let tx = spend_nullifiers_tx(self, builder.evm_mut(), spent_nullifier_hashes).map_err(
-                |e| {
-                    error!(target: "payload_builder", %e, "failed to build spend nullifiers transaction");
-                    PayloadBuilderError::Other(e.into())
-                },
-            )?;
-            let gas_used = builder.execute_transaction(tx.clone()).map_err(
-                |e: BlockExecutionError| {
-                    error!(target: "payload_builder", %e, "spend nullifiers transaction failed");
-                    PayloadBuilderError::evm(e)
-                },
-            )?;
-
-            self.commit_changes(info, base_fee, gas_used, tx);
-        }
-
-        if !invalid_txs.is_empty() {
-            pool.remove_transactions(invalid_txs);
-        }
-
-        Ok(None)
-    }
-
-    /// After computing the execution result and state we can commit changes to the database
-    fn commit_changes(
-        &self,
-        info: &mut ExecutionInfo,
-        base_fee: u64,
-        gas_used: u64,
-        tx: Recovered<OpTransactionSigned>,
-    ) {
-        // add gas used by the transaction to cumulative gas used, before creating the
-        // receipt
-        info.cumulative_gas_used += gas_used;
-        info.cumulative_da_bytes_used += tx.length() as u64;
-
-        // update add to total fees
-        let miner_fee = tx
-            .effective_tip_per_gas(base_fee)
-            .expect("fee is always valid; execution succeeded");
-        info.total_fees += U256::from(miner_fee) * U256::from(gas_used);
-    }
-
-    /// Prepares [`BlockBuilder`] for the payload. This will configure the underlying EVM with [`PBHCallTracer`],
-    /// but disable it by default. It will get enabled by [`WorldChainPayloadBuilderCtx::execute_best_transactions`].
-    fn block_builder<'a, DB: Database>(
-        &'a self,
-        db: &'a mut State<DB>,
-    ) -> Result<
-        impl BlockBuilder<
-            Primitives = OpPrimitives,
-            Executor: BlockExecutor<Evm = OpEvm<&'a mut State<DB>, NoOpInspector>>,
-        >,
-        PayloadBuilderError,
-    > {
-        // Prepare attributes for next block environment.
-        let attributes = OpNextBlockEnvAttributes {
-            timestamp: self.inner.attributes().timestamp(),
-            suggested_fee_recipient: self.inner.attributes().suggested_fee_recipient(),
-            prev_randao: self.inner.attributes().prev_randao(),
-            gas_limit: self
-                .inner
-                .attributes()
-                .gas_limit
-                .unwrap_or(self.inner.parent().gas_limit),
-            parent_beacon_block_root: self.inner.attributes().parent_beacon_block_root(),
-            extra_data: self.inner.extra_data()?,
-        };
-
-        // Prepare EVM environment.
-        let evm_env = self
-            .inner
-            .evm_config
-            .next_evm_env(self.inner.parent(), &attributes)
-            .map_err(PayloadBuilderError::other)?;
-
-        // Prepare EVM.
-        let evm = self.inner.evm_config.evm_with_env(db, evm_env);
-
-        // Prepare block execution context.
-        let execution_ctx = self
-            .inner
-            .evm_config
-            .context_for_next_block(self.inner.parent(), attributes);
-
-        // Prepare block builder.
-        Ok(self
-            .inner
-            .evm_config
-            .create_block_builder(evm, self.inner.parent(), execution_ctx))
-    }
-}
-
-pub const COLD_SSTORE_GAS: u64 = 20000;
-pub const FIXED_GAS: u64 = 100_000;
-
-pub const fn dyn_gas_limit(len: u64) -> u64 {
-    FIXED_GAS + len * COLD_SSTORE_GAS
-}
-
-pub fn spend_nullifiers_tx<DB, I, Client>(
-    ctx: &WorldChainPayloadBuilderCtx<Client>,
-    evm: &mut OpEvm<DB, I>,
-    nullifier_hashes: HashSet<Field>,
-) -> eyre::Result<Recovered<OpTransactionSigned>>
-where
-    I: Inspector<OpContext<DB>>,
-    DB: revm::Database + revm::DatabaseCommit,
-    <DB as revm::Database>::Error: std::fmt::Debug + Send + Sync + derive_more::Error + 'static,
-{
-    let nonce = evm
-        .db_mut()
-        .basic(ctx.builder_private_key.address())?
-        .unwrap_or_default()
-        .nonce;
-
-    let mut tx = OpTransactionRequest::default()
-        .nonce(nonce)
-        .gas_limit(dyn_gas_limit(nullifier_hashes.len() as u64))
-        .max_priority_fee_per_gas(evm.ctx().block.basefee.into())
-        .max_fee_per_gas(evm.ctx().block.basefee.into())
-        .with_chain_id(evm.ctx().cfg.chain_id)
-        .with_call(&spendNullifierHashesCall {
-            _nullifierHashes: nullifier_hashes.into_iter().collect(),
-        })
-        .to(ctx.pbh_entry_point)
-        .build_typed_tx()
-        .map_err(|e| eyre!("{:?}", e))?;
-
-    let signature = ctx.builder_private_key.sign_transaction_sync(&mut tx)?;
-    let signed: OpTransactionSigned = tx.into_signed(signature).into();
-    Ok(signed.into_recovered_unchecked()?)
 }
