+++ crates/world/payload/src/lib.rs
@@ -1,2 +1,3 @@
-#![cfg_attr(not(test), warn(unused_crate_dependencies))]
 pub mod builder;
+
+pub mod ctx;
